Small functional shuffle
