+++ scripts/metadata_comparison/test/lib/helper_functions.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+def read_resource(filename):
+    path = f'test/resources/{filename}'
+    with open(path, 'r') as file:
+        data = file.read()
+    return data
